Convert branches to commits for comparisons
Sometimes the symmetric git log checks fail if one of the commits being
used is a branch.  This slightly hackish solution does a naive check of
commit string length to see if it's an actual commit (40 is unlikely
for a branch name length, but taking a slight chance here), and if not,
a repo command is run to determine the latest commit on that branch.

Tested and seems to be doing the correct thing.  A better solution may
be forthcoming.

Change-Id: I8702b53bd5aa102070c09c4294c813960899815c
Reviewed-on: http://review.couchbase.org/83760
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Kenneth Lareau <kenneth.lareau@couchbase.com>