CBD-2423: Move configuration data out of repo_upload
Part of the process in removing dynamic data from application builds,
moved the JSON data files into product-metadata and added a '-D' option
to set the path to the files for the program to use.  An additional
change related to that was to throw an exception caught by the top-level
for known possible failures of reading the JSON files.

Change-Id: I77a64f6c0ea3eb1b0a1fd9848cb8e6d9aceed357
Reviewed-on: http://review.couchbase.org/91869
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Kenneth Lareau <kenneth.lareau@couchbase.com>