CBD-2472: Update generate_filelist() to handle multi-packages
With the addition of sync-gateway to the set of prodcuts being checked,
given there is the need to be able to recognize multiple packages for
a single product, the generate_filelist() method needed to be updated
to handle this.

The new format puts the specific data into a new 'packages' key that
has each package as a separate entity with its own information on how
to generate the filelistings for it.

Change-Id: Id6ba560f3924b8e72219d6362e4618619eae7aa2
Reviewed-on: http://review.couchbase.org/94118
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Kenneth Lareau <kenneth.lareau@couchbase.com>