Better product/manifest_path handling in build_from_manifest
With introduction of more top-level directory product 'groups', a more
general way to determine the split between product and manifest path
was needed.  The new method recursively searches up the directory path
until a 'product-config.json' file is found.

Note that this now requires every product to have this file, which may
not fully be the case; for those that don't, one will need to be made.

Change-Id: I85eb1b0a74f3fb761732581a17a9ad3c6704968d
Reviewed-on: http://review.couchbase.org/101289
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Kenneth Lareau <kenneth.lareau@couchbase.com>