Better product handling in build_from_manifest
With introduction of more top-level directory product 'groups', a more
general way to determine the split between product was needed.  The new
method recursively searches up the directory path until a 'product-
config.json' file is found; the manifest_path, having been a temporary
variable, is no longer needed.

Note that this now requires every product to have this file, which may
not fully be the case; for those that don't, one will need to be made.

Change-Id: I680ba9cce55f10040431b6a44eff35022125061e
Reviewed-on: http://review.couchbase.org/101362
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Kenneth Lareau <kenneth.lareau@couchbase.com>