Add new properties parameters for build_from_manifest
Adding the following new parameters to the build properties files:
  - PROD_NAME
  - PRODUCT_PATH

PROD_NAME is the name of just the product (minus any scoping prefix)
and PRODUCT_PATH - formerly MANIFEST_PATH - is the path in the manifest
repo that the product's main information resides in.

Both will be used for passing down from the build_from_manifest job
in Jenkins to downstream jobs that need the information.

Change-Id: I6453d30dd5954abf0c1a386842375109139e68e0
Reviewed-on: http://review.couchbase.org/101767
Reviewed-by: Kim Trang <kim.trang@couchbase.com>
Tested-by: Kenneth Lareau <kenneth.lareau@couchbase.com>