CBD-2816: Initial SDK source-download scripts
Includes scripts and Black Duck configuration for:
  - couchbase-sdk-c (libcouchbase)
  - couchbase-sdk-net (.NET)
  - couchbase-sdk-nodejs
  - couchbase-sdk-java
  - couchbase-sdk-python

These scripts are best-guess based on earlier conversations with SDK
engineers and simply looking at the code. It is unclear whether the scan
results for these packages are accurate. Python in particular is known not
to be.

This commit also includes a couple fixes for the general detect-scan scripts.

Change-Id: Ida4ff4ddabfc9a15fb9eee66b8e9d7d31ddc6278
Reviewed-on: http://review.couchbase.org/118992
Reviewed-by: Ming Ho <ming.ho@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>