CBD-3642: Avoid Black Duck scan errors from component name changes
If Black Duck changes components names, when we start scanning a new
version of the product the existing black-duck-manifest.yaml files
will cause errors. This is because we look up "new" components in
the Black Duck database using fuzzy matches, but then later compare
the manifest component names with the database names using exact
matches.

Now we will warn in such a situation and skip the fuzzy match, with
a note to tell users to update the black-duck-manifest.yaml files with
an exact match when necessary.

Change-Id: Ia3c344442323987c129fff217a036a12fd821a68
Reviewed-on: http://review.couchbase.org/c/build-tools/+/136245
Tested-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>