CBD-3879: Upload k8s artifacts to S3 in product-specific subdirs
Historically all k8s products were uploaded to
packages.couchbase.com/kubernetes. This gets confusing as more products
are added with differing versioning schemes. It also is inconsistent
with the directory structure on releases.build.couchbase.com which is in
product-named subdirectories like all other Couchbase products.

Note: This change does not touch the older "operator-release.sh" script.
That script is no longer used for Operator >= 2.1, and is only still in
this repository in case of an MP release in future.

Change-Id: I6bce90b14564ee28bf26a57c401c26d658463040
Reviewed-on: http://review.couchbase.org/c/build-tools/+/152066
Reviewed-by: Blair Watt <blair.watt@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>