CBD-4185: Handle a situation where the first module_projects project is not root
Instead of running the update SHA function once, run it for each repo.  This should not affect the existing projects because it is only additive.  The same set of subprojects will be discovered and the same precalculated SHAs will be used.  The only slight redundancy is that another iteration of the loop will be performed to look at the projects that have already been looked at and discover that they are already all correct.

Change-Id: Id30bfebd85ad71d56ec1a1b6f8629f4770f6cba4
Reviewed-on: http://review.couchbase.org/c/build-tools/+/156909
Reviewed-by: Ming Ho <ming.ho@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Jim Borden <jim.borden@couchbase.com>