CBD-4271: update-manual-manifest can take Black Duck IDs
Manual Black Duck manifests may now specify "bd-id" fields for
components, in which case update-manual-manifest will use that ID. This
allows us to avoid troubles where Black Duck has difficulty finding the
right version of, say, v8 when starting scans for a new version of
Server. If bd-id is not specified, it will revert to the old behaviour
of (a) trusting the ID of an existing manual component with exactly the
same name, or else (b) using the Black Duck REST API to search for a
component with the same name and at least one matching version.

Implementing this required a fairly thorough overhaul of the internal
data structures, which also changed the program flow somewhat. In
particular, all unknown BD IDs are looked up first, rather than during
diffing. This has the nice side effect of preventing Black Duck from
ending up in a partially-updated state when the searching heuristic
fails.

I also discovered some new REST APIs to make the searching for
components and versions faster and likely more reliable.

In general the program is more robust and better documented.

Change-Id: I0461d154443e5b826f8050fb21622987833138b2
Reviewed-on: http://review.couchbase.org/c/build-tools/+/159946
Reviewed-by: Blair Watt <blair.watt@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>