CBD-4359: Add erlang cbdeps v2, including Windows!
The Unix build script is basically unchanged from the cbdeps v1 script
from tlm/deps/packages/erlang.

The Windows script is massively shortened, as the new WSL-based build
method they've devised requires a lot fewer wacky workarounds. I was
able to update our OpenSSL packaging on Windows to include static
libraries, so that this Erlang build could use that rather than
requiring to download a third-party OpenSSL build. Also, we no longer
build an NSIS-based installer just to throw it away; I've updated the
packaging to include Erlang's "Install.exe" which is executed as part of
the cbdeps package install, the same as it is on Unix.

I extended build-one-cbdep and build-one-cbdep.bat to always download
the latest cbdep tool and put it on the PATH, to make it easier for
cbdeps builds to grab their own dependencies.

Finally, I slightly cleaned up the way build-one-cbdep.bat and
win32-environment.bat handle MSVC versions. And going forward, new
cbdeps packages will simply be named "windows" rather than eg.
"windows_msvc2017". The choice of compiler matters when choosing Jenkins
agents to build on, but not when downloading dependency packages, so
long as we bundle in the latest version of the MSVC redistributables
package (which I've done in voltron, and updated the build-infra
playbooks to ensure it's available).

Change-Id: Iec846381360c754352e79b5f73da3be04de817d0
Reviewed-on: http://review.couchbase.org/c/build-tools/+/163157
Reviewed-by: Blair Watt <blair.watt@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>