CBD-4440: Enable customizing Go version for operator scans
Introduce optional "scan-environment.sh" file in blackduck/<product>
which, if it exists, will be invoked and the stdout from it will be
eval'd by the scanning script. Intended for setting up things that
should be added to the environment for scanning jobs. In the case of
couchbase-operator, ensures that a new version of Go is on the PATH.

Change-Id: I8fca0efa7290173a5a44c711710089073e308a00
Reviewed-on: http://review.couchbase.org/c/build-tools/+/164713
Tested-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>