MB-49865: Allow toy builds to be based on historic build manifests
This is a generally useful feature, but we are implementing it in
particular for the Server weekly staging run, where we need to create a
Server build that is "the same" as the weekly build, but with a higher
build number (for the purposes of upgrade testing). We are by
convention reserving the build numbers 30000-49999 for these upgrade
builds.

Change-Id: Ie0bb00fa3e3f6c229d28576e598bcc49661269c6
Reviewed-on: https://review.couchbase.org/c/build-tools/+/168740
Reviewed-by: Blair Watt <blair.watt@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>