CBD-4541: Determine Analytics components from their BOMs
Analytics now creates a set of BOM files in the form
groupId:artifactId:version. At Black Duck scan time, we convert those
to a specialized set of pom.xml files, and have BD scan only those.
That results in BD reporting precisely the components reported by
Analytics.

Note that to do this, we need to build Analytics as part of the BD
scan, which takes some number of minutes. However, this allows us to
NOT download or scan any .jar files, which saves us considerable
BD scanning space.

This change downloads JDK 11 for building Analytics, and also uses
scan-environment.sh to instruct BD Detect to use that JDK for its
invocations of maven, to avoid any JDK-version-related problems.

Change-Id: If7399b61f098ba17dd073a277659640377fde748
Reviewed-on: https://review.couchbase.org/c/build-tools/+/169697
Reviewed-by: Blair Watt <blair.watt@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>