MB-51030: Patch curl to allow specifying CABUNDLE via environment
Curl bakes in the path to the OS's CA bundle at build time. Since we
want to run the same libcurl bits on multiple OSes, we need a way to
modify this based on the target system. The "curl" binary supports
overriding this value at runtime using the environment variable
CURL_CA_BUNDLE. Here we patch libcurl to have the same ability. This
can be enabled using --with-ca-bundle=env when using the configure
script, or with the CMake option CURL_CA_BUNDLE=env.

Change-Id: I76612a9a3c68e2b21cafb907aa512d6c5c5f1ef3
Reviewed-on: https://review.couchbase.org/c/build-tools/+/171463
Reviewed-by: Ming Ho <ming.ho@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>