Fix couchbase-sdk-java and couchbase-sdk-scala BD scans
Need to prune java-fit-performer before scanning. It is a test framework
so it doesn't need to be scanned. Normally it would be skipped due to
detect.maven.included.modules not including it, but Black Duck Detect
still needs to run "mvn dependency:tree" on the pom.xml to determine the
module name for that to work. Unfortunately java-fit-performer/pom.xml
throws an error at that stage unless certain additional libraries are
manually installed.

Change-Id: I211b2a4380b5938efb85a980db387eae5fb38b44
Reviewed-on: https://review.couchbase.org/c/build-tools/+/173857
Tested-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Ming Ho <ming.ho@couchbase.com>