CBD-4768: Fix BD version matching once and for all
The Black Duck Knowledgebase is sometimes inconsistent with the version
names it assigns to components, which leads to spurious problems where
we compare versions and get incorrect results. There have been several
attempts to rectify this over time, but they've all had problems such as
only catching inconsistencies from BD but not in manifests, or only for
existing components but not when looking up new ones. This fix should
finally make it completely consistent.

We canonicalize version names for:

 - Erlang: no "OTP-" prefix
 - Golang: no "go" prefix

This logic is now isolated into a single function canonicalize_version,
so we can handle additional components in future as necessary. We call
this new function:

 - in _load_manual_components(), to ensure our view of the versions
   currently in Black Duck (comp_map) uses canonical versions
 - in add_manifest(), to ensure our internal representation of any
   manifest YAML files uses canonical versions
 - in find_version_for_component(), which is used when looking up new
   versions in the Black Duck Knowledgebase (either because we are
   adding a new version, or because we're reading an older manifest YAML
   that doesn't use bd-id) so we can match regardless of what version
   name form Black Duck returns from the search API

In short, all in-memory version name representations will be
canonicalized, meaning that we can safely compare them directly, and
that dictdiffer can do its job correctly.

Change-Id: I00be7bc739fbd6a959a4a0b72420164a9addb98d
Reviewed-on: https://review.couchbase.org/c/build-tools/+/175232
Reviewed-by: Blair Watt <blair.watt@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>