Modify logic for finding <annotation> elements in manifest
Due to V8, we have a manifest with two "build" elements. Therefore we
need to make the search for "build" project more stringent. Now we
search for the actual "VERSION" annotation itself directly, and use that
as an anchor for adding additional annotations in the build manifest.
Due to this, we must make the VERSION annotation strictly required,
where previously it was technically optional although nobody used that.
Since I was making this change anyway, I went ahead and made the
'release' entry in product-config.json option, defaulting to VERSION.
This will be useful for projects that always have VERSION and release
set the same.

Change-Id: I0e4b4eb6fdfd919cc6ef81b82e9ad251d7f630cb
Reviewed-on: https://review.couchbase.org/c/build-tools/+/179603
Reviewed-by: Blair Watt <blair.watt@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>