CBD-5177: Update preflight image certification to new RHCC specs
The new quay.io-based RHCC behaves quite differently than the old. Now
we can upload images with whatever tags we want, including multiple tags
for the same image and (wonder of wonders) moving one tag to another.
Also, auto-publish works, so whatever tags we push will be available
automatically after the preflight certification results are uploaded.

I have revamped the old preflight_certify.sh to a new
rhcc-certify-and-publish.sh script, which handles both uploading an
image to RHCC (with multiple tags) and running preflight on it. It also
integrates this with the publish-k8s-images.sh script, which now can
fully handle Docker Hub and RHCC release activities. The old
publish-rhcc-image.sh script in the couchbase-partners/redhat-openshift
repository will be retired entirely as rhcc-certify-and-publish.sh can
handle all those duties.

Also, this new script can run on "buildx" agents, as it no longer
requires the manually-maintained /home/couchbase/openshift directory on
the older rhel75 agent. This was achieved via a new Docker Swarm secret
which captures all the necessary RHCC upload keys in a JSON file.

It is unfortunately still necessary to log in to the RHCC UI after
uploading an image this way, and selecting "Sync Tags" from the
three-dot menu. Otherwise, only the first tag specified to the publish
script will show up in the publicly-visible Catalog webpage (although
all tags will be available for `docker pull`). I tried several ways to
eliminate this step, but the site does not currently allow for it. Maybe
someday.

Change-Id: I12c5d9e5d8b6bdef2c6152ae2e857572b763544b
Reviewed-on: https://review.couchbase.org/c/build-tools/+/181717
Reviewed-by: Ming Ho <ming.ho@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>