Hack script to clean Windows registry
If Server install or uninstall fails, or a buggy installer is used,
the Windows registry can be left with a bunch of cruft in it. These
stale entries can cause future install/uninstall actions to operate
incorrectly. This script assumes that Server has be uninstalled as
thoroughly as it can be ("CouchbaseServer" Windows service stopped
and removed; C:\Program Files\Couchbase\Server removed; registry
hacks done to remove Couchbase Server from the Add/Remove Programs
settings dialog), and simply deletes any entries in the registry
which refer to files and directories that no longer exist.

Change-Id: Ia60b01b03321498499ac693e8bbbfc017a3ea2cc
Reviewed-on: https://review.couchbase.org/c/build-tools/+/185669
Tested-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>