MB-55204: Don't rebuild just for product-texts changes
This allows us to keep all manifests pointing to the 'master' git branch
of product-texts (so license changes will be ingested automatically)
without worrying that pushing a change to product-texts will trigger
dozens of builds of old product versions. It does mean that we need to
explicitly remember to trigger a new build if there's a license change
right at the end of a product version cycle.

Also fix --project-skip option for lock-to-sha.py, and add product-texts
to the set of default projects not to lock.

Change-Id: Ibc6c8ffa1cf694cfe9761b680ff3b9982e65f6c6
Reviewed-on: https://review.couchbase.org/c/build-tools/+/186188
Reviewed-by: Blair Watt <blair.watt@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>