CBD-5309: Don't rebuild just for golang changes
Since the "golang" repository is shared by a number of product versions
and we want them all pointing to 'master', but we don't necessarily want
a change in the golang repo to trigger builds for all those versions.

Also: reformatted the check for project names to make it more readable;
switch the shebang line to just say "python3" rather than hardcoding
"python3.6" (which was a historical workaround for a deficiency in the
zz-lightweight builders, long since corrected).

Change-Id: Iceb078f50d05b6309f1d9a6bf4dc589f9f106f77
Reviewed-on: https://review.couchbase.org/c/build-tools/+/186254
Tested-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Blair Watt <blair.watt@couchbase.com>