Prune Black Duck scans for released connector versions
According to Ceej, we'll still get ongoing vulnerability
reports for the old versions; scanning just updates the
list of dependencies associated with each version.

Dependencies are set in stone when a version is released,
so scan definitions for released versions don't need to
stick around forever. We only need to keep them long enough
for the final release to be scanned.

Change-Id: If8e3048d4bac4426e4a6638c1c684cdd6c305f5b
Reviewed-on: https://review.couchbase.org/c/build-tools/+/186743
Tested-by: David Nault <david.nault@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>