CBD-3676: Republishing K8S Docker images with updated base images
Introduce new couchbase-k8s-microservice-republish-all-images.sh to
match corresponding Jenkins job. This scans all released versions of
K8S products; identifies the available GA versions; rebuilds the
Docker Hub and RHCC images as appropriate; and if the images have now
changed, publishes them.

Changed internal build-k8s-images.sh and publish-k8s-images.sh to
support rebuilding/republishing. They now explicitly save the build
cache to build-docker (under the cb-vanilla-buildcache and
cb-rhcc-buildcache namespaces) to ensure that the exact same image is
created when rebuilding when the base image has not changed. Also,
build-k8s-images.sh has a new -P option which calls publish-k8s-images
immediately.

Note: we no longer pass the build arg OS_BUILD to the Docker image
builds, because that trashes caching which results in the locally-built
image always appearing "new" compared to the published image. This build
arg was only used to set certain Docker image labels on RHCC. It is far
from clear that those image labels were ever necessary or correct.

Changed rhcc-certify-and-publish.sh to accept an ARCH argument so that
it publishes amd64 and arm64 images independently. This was necessary
for cases where the amd64 and arm64 images needed to be rebuilt at
different times.

Changed release process to always save the -image file on the internal
release mirror - previously only saved for couchbase-operator as it is a
shipped artifact for that product - as it is necessary to use for
rebuilding the images. It will still only upload this file to S3 for
couchbase-operator.

Added logic to publish-k8s-images.sh to avoid uploading at all if the
image has not changed. This was necessary because on Docker Hub, pushing
the identical image still modified the timestamps shown on the website.
On RHCC, pushing and then attempting to preflight-check the identical
image would raise "cannot preflight already-published images" errors.

On RHCC, ensure that we preflight exactly the "X.Y.Z-b" image as this
tag is guaranteed to be different per version. This avoids the same
"cannot preflight already-published images" error if we attempt to
preflight check a newly-pushed "X.Y.Z" image.

Adds several generally-useful functions to utilities/shell-utils.sh for
investigating images on Docker Registries, and a new run_untraced
function for suppressing bash's xtrace for a single command to avoid
leaking sensitive information.

Change-Id: Ia76f33d562071be71936dc17f1089db659e6541b
Reviewed-on: https://review.couchbase.org/c/build-tools/+/188351
Reviewed-by: Blair Watt <blair.watt@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>