MB-56691: Ensure erl TERMINFO_DIRS is correct
We use libtinfo from cbpy for erlang, however the library in
in question has a terminfo path hardcoded which only exists on
the upstream conda build systems, which is causing issues when
running `couchbase-cli server-eshell`.

This change modifies the `erl` script template to ensure we are
overriding the missing path via the TERMINFO_DIRS environment
variable.

Change-Id: I6c90fc97d11eba040a5a42d32c116e67d5579e1b
Reviewed-on: https://review.couchbase.org/c/build-tools/+/190839
Reviewed-by: Ming Ho <ming.ho@couchbase.com>
Tested-by: Blair Watt <blair.watt@couchbase.com>