CBD-5433: Switch cbdeps 'profile' version numbering convention
Previously building a cbdeps package with PROFILE set to anything other
than the default of "server" changed BLD_NUM to be ${PROFILE}${BLD_NUM},
eg "sdk5". This was always a little weird, but is extra bad for curl
as we wanted to have an "openssl111" profile. This resulted in build
numbers like "openssl1113" which is impossible to parse.

Switch to ${BLD_NUM}_${PROFILE}, eg. 3_openssl111. This is clear,
and still allows combined version numbers like "7.88.1-3_openssl111"
to be split on - if necessary.

Change-Id: I973c5e8d165329db8675bfb554063728936a70de
Reviewed-on: https://review.couchbase.org/c/build-tools/+/191319
Tested-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>