CBD-5652: Fix Black Duck Goldfish scans
Currently, any vanilla scans are running with the default depth
of 1 and firing in src/ so do not find any of the source code.
This change adds detect-config.jsons for kafka-connect-dynamodb
and couchbase-goldfish-nebula to ensure their scans descend into
the source, and adds prune scripts to remove files in src/ which
should not be included in the scan.

Change-Id: I2bc21145970734d1e77bc45ef53e9fb46cd3c7b7
Reviewed-on: https://review.couchbase.org/c/build-tools/+/198580
Reviewed-by: Ming Ho <ming.ho@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Blair Watt <blair.watt@couchbase.com>