CBD-5666: Add "license-approved" field in black-duck-manifest.yaml files
If this field is specified for a component, update-manual-manifest will
set the "reviewed" status for all versions of that component in the
product-version accordingly.

Empirically, setting this status by hand in the Black Duck UI sometimes
gets "lost", possibly by later runs of this script. Having it set
programmatically as part of the manual manifest should prevent this.

Also, don't report any components beginning with "Couchbase" (case-
insensitive) as "Suspect Licenses"; and fix a Python warning.

Change-Id: Idecb66edf749064c8951e8ab36abd32a428b4bbc
Reviewed-on: https://review.couchbase.org/c/build-tools/+/199311
Tested-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Blair Watt <blair.watt@couchbase.com>