MB-59595: Build openblas cbdeps package with TARGET options
By default, OpenBLAS builds for the instruction set of the CPU it is
built on. We need to configure it for an older TARGET so that the
resulting lib will work on any CPU we support. While we're fixing that,
also enable the DYNAMIC_ARCH feature which allows it to determine the
instruction set at runtime, and a few other build option cleanups. So
far the DYNAMIC_ARCH feature is only enabled on x86_64; our compiler on
aarch64 doesn't seem to support all of the ARMV8 architectures that
OpenBLAS wants (in particular it complains about NEOVERSEV1).

Change-Id: Iecb923b08a423fb1f649579a59d684188810d2ae
Reviewed-on: https://review.couchbase.org/c/build-tools/+/200755
Tested-by: Ming Ho <ming.ho@couchbase.com>
Reviewed-by: Abhi Dangeti <abhinav@couchbase.com>