VULN-3333 handle a special case of re-opening VULN
Sometimes, CVE is removed after discovering it doesn't apply
to a BD component.  As a result, our script closes the VULN
ticket and removed values in BD_CVES field.  This is fine.  However,
if a new CVE is later found for the BD component, our script will
re-open the ticket, we need to make sure it take into consideration
that BD_CVES could be empty.

-Ming

Change-Id: I5c117d112417a458b8d76cfb5eeca6d9abc57424
Reviewed-on: https://review.couchbase.org/c/build-tools/+/203401
Tested-by: Ming Ho <ming.ho@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>