Safer flatten-project-version
Read the existing signature scan components from the most recent
saved components.csv in product-metadata, rather than just keeping
them in memory while deleting the signature scans.

Don't treat "manual component already exist" as a fatal error, as this
seems to occur regularly. Also, in this way, this script is re-entrant
and can be run again on an already-flattened project-version.

Change-Id: I6273f65aa3e7687b9fe6ef02b18b5d568d2b569a
Reviewed-on: https://review.couchbase.org/c/build-tools/+/203793
Reviewed-by: Blair Watt <blair.watt@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>