Add check to detect need to update bd_aliases.yaml
If Black Duck changes a component ID in their Knowledgebase, one likely
failure for our scripting is that searching for a version of the old
component ID will return a component-version URL with the new component
ID. We will then fail attempting to add it to the BOM because it will
actually already exist in the BOM.

It would be challenging to auto-correct for this situation, and I'm not
sure the "old" component IDs exist forever in the Knowledgebase anyway.
However we can at least identify the situation and provide a useful
error message, rather than a mysterious "error 412" when adding the
changed component-version URL.

Also add bd-id-aliases entry for abseil, which lead to this.

Change-Id: Ib6689bcdd8fe150176cc3e0c6a01f81f41dd20de
Reviewed-on: https://review.couchbase.org/c/build-tools/+/204068
Reviewed-by: Ming Ho <ming.ho@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>