Use npm from blackduck-scan VM
couchbase-cloud now requires a newer npm than was specified in this
script. Newer versions from cbdep don't run on the BD scan VM (requires
a newer glibc), so I built npm 20 from source on the VM after installing
gcc 8. Use that for now, until we can update the VM to a newer OS or
move it to a Docker container.

Change-Id: If12c33a5b2fb25d76b9d4011a9d458a9b5eb81f7
Reviewed-on: https://review.couchbase.org/c/build-tools/+/204542
Tested-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Ming Ho <ming.ho@couchbase.com>