MB-60960: Convert cbpy to cbdeps 2.0
This code is virtually identical to the original cbdeps 1.0 version at
https://github.com/couchbase/tlm/tree/fa220f9cf953e11952371cb053185368de8b18f7/deps/packages/cbpy
with the following differences:

- cbpy_unix.sh and cbpy_windows.bat accept slightly different args,
  based on cbdeps 2.0 build-one-cbdep scripts
- black-duck-manifest.yaml moved to package/ subdir, so it is included
  in the final cbdeps tgz files
- CMakeLists.txt removed (no longer needed)
- MINIFORGE_VERSION setting moved to cbpy_unix.sh/cbpy_windows.bat
- verify-black-duck-manifest.py --tlm_dir arg changed to --src_dir
- README.md updated with relevant process changes
- Apple codesign step in cbpy_unix.sh removed - it was actually unused
  for the past year, and shouldn't be necessary as Server's own
  codesigning process will also sign Python

MB-60960 was actually about updating py/cryptography. Since that package
isn't pinned to a version in cb-dependencies.txt, building this package
now automatically picks up the newer 42.0.5 version.

I've pinned python-snappy to < 0.7.0, because 0.7.0 switched the
implementation to the entirely new "cramjam" Rust-based implementation
and no longer actually uses Snappy. We will want to upgrade at some
point, but we need to verify with Dev that the behaviour remains
unchanged.

Change-Id: I3eb0818f28158a9f8d2a7fdf11d213db79b2aee5
Reviewed-on: https://review.couchbase.org/c/build-tools/+/207733
Reviewed-by: Ming Ho <ming.ho@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>