CBD-5848, reopen VULN if status in "Not Applicable" when a new CVE comes in.
If a new CVE comes in, we should make sure appropriate VULN is reopened.
"Not Applicable" means the original list of CVEs doesn't apply.  When a new CVE
appears, we need to reopen the ticket to re-evaluate.

Also removed "low_severity" from workflow as it is not been used at all.
Updated workflow transition IDs due to workflow changes.

-Ming

Change-Id: I74f86003b2098936e0ee2637e1d39ac410c4d30b
Reviewed-on: https://review.couchbase.org/c/build-tools/+/207851
Tested-by: Ming Ho <ming.ho@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>