Automatically release couchbase-operator sub-projects
This will allow a single run of the couchbase-k8s-microservice-release
job for couchbase-operator to automatically release the Docker/RHCC
images for the "sub-projects" couchbase-admission-controller and
couchbase-operator-certification. All other release steps - triggering
update_manifest_released and release_binaries_to_s3, etc - are only done
for couchbase-operator, which is the way it's always been.

Change-Id: I5f73729d43160bca73451b0e328ebc160d29f46e
Reviewed-on: https://review.couchbase.org/c/build-tools/+/208987
Reviewed-by: Ming Ho <ming.ho@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>