MB-61410: Fixes for jemalloc_noprefix on Windows
The .dll and .lib files created by a jemalloc Windows builds are named
after the basename of the .vcxproj file used to build them, which is
called `jemalloc.vcxproject` in the upstream source. When building with
--with-install-suffix, we initially just copied the resulting
jemalloc.dll and jemalloc.lib files to jemalloc_noprefix.dll and
jemalloc_noprefix.lib. However the .lib file also has internal
references to the name of the dll. These were still naming jemalloc.dll,
which led to the wrong jemalloc library being used at runtime.

Instead, now we rename the jemalloc.vcxproj file itself prior to
building it.

Change-Id: I636713201382ddd08c2dae3361ba81f2ad87a1cd
Reviewed-on: https://review.couchbase.org/c/build-tools/+/209356
Tested-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Blair Watt <blair.watt@couchbase.com>