Correct BD component for grpc
It seems Synopsys has been changing/splitting component IDs again. The ID we
used to list as canonical for grpc (4c1ff00) is now "Grpc.Tools" from Nuget;
whereas the ID we've had in Black Duck manifest files (a38e1dc) is now
"grpc/grpc-dotnet". The canonical version now appears to be d720ec7, and this
is the ID that is Black Duck Detect uses for grpc as referenced by maven as
well.

Change-Id: Ifc9ca98a2d2fc22c827f47a5f326b4d3dde3f961
Reviewed-on: https://review.couchbase.org/c/build-tools/+/210606
Reviewed-by: Ming Ho <ming.ho@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>