Support for jenkins_job_parameters in product-json.config
Introduce new per-manifest config option "jenkins_job_parameters", which
is a JSON object. The key/value pairs in that object will be passed as
additional parameters to the Jenkins build job.

First usage of this functionality are faiss builds (both production and
toy), which want to pass AUTO_PUBLISH=true to cbdeps-build. Since this
is now done via this general-purpose mechanism, we can remove the
special-case code in scan-manifests which was previously populating that
parameter for toy builds.

Change-Id: I0cc037b5dc9dab21b577bb65bca2ea4663d2526b
Reviewed-on: https://review.couchbase.org/c/build-tools/+/210737
Tested-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Blair Watt <blair.watt@couchbase.com>