CBD-6020: Black Duck: Check alt-canonical versions for fallback versions
We already have a concept of "alternate canonical versions" to protect
us when eg. ca-certificates versions in the Knowledgebase switch from
2024.02.02 to 2024.2.2. We also have the concept of "fallback versions"
to protect us from cases where we depend on a version that the
Knowledgebase doesn't have yet. However, we hadn't joined the two - when
using a fallback version, we assumed it was valid. That doesn't work
when the fallback version itself changes.

Now when we decide to try a fallback version, we also canonicalize it
and save alt-canonical versions.

Change-Id: I2c458fa21981f706ab87f206bcb0a7c74c773f2f
Reviewed-on: https://review.couchbase.org/c/build-tools/+/215012
Tested-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Ming Ho <ming.ho@couchbase.com>