CBD-6020: More work with canonical fallback versions
The function find_canonical_component_version() takes a canonicalized
version and looks it up in the Knowledgebase. It will try
"alt-canonical" versions also; that's the point of the function, to find
the component_version_url for the *canonical* version name, even if the
Knowledgebase calls it something different.

In the past, this functioned returned the actual discovered version name
from the Knowledgebase when it discovered an alt-canonical version.
However, it is vital that `bom_comp_map` and `manifest` only ever
contain canonical version names. Therefore it is not useful to return
this discovered version, and can lead to confusion. Remove that return
value.

For the same reason, change fallback_version_if_necessary() to always
return only a canonical version, even if that was a fallback version.

Change-Id: I2435395d0631e8e65ad30da1640f506ac6a0121b
Reviewed-on: https://review.couchbase.org/c/build-tools/+/215247
Tested-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>