Add possibility to skip certain safety checks when publishing to RHCC
I recently added checks to ensure that we didn't upload and
already-existing image (checked by manifest SHA) or an already-existing
image tag to the redhat-isv-containers site. This was necessary to avoid
a number of problems with tags and publishing.

This change adds a -f argument to rhcc-certify-and-publish.sh which
bypasses those checks (I've also added a FORCE argument to the
publish-rhcc-image Jenkins job). This is necessary when an initial image
upload fails at preflight time but we need to try again, for example
when our version of preflight gets too old and RHCC rejects the report
upload. This flag should only be used when re-attempting an upload of an
image and/or tag that has never been *published* on RHCC.

Also, update preflight to 1.10.0.

Change-Id: If45ab105f94b558b82540b1e9f34082b296c9829
Reviewed-on: https://review.couchbase.org/c/build-tools/+/215306
Reviewed-by: Blair Watt <blair.watt@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>