CBD-6104: Commit .python-version files for uv scripts
`pyproject.toml` declares a range of Python versions that the project
is compatible with. `.python-version` declares a specific Python
version that should be used when running the scripts in the project.
They overlap, but they're not quite the same thing. Especially since
these scripts run in CI environments, it's probably best to use both
so we're explicit about what will run.

Change-Id: I012d668a5a76ca2498840f088414420239a824b1
Reviewed-on: https://review.couchbase.org/c/build-tools/+/219330
Well-Formed: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Ming Ho <ming.ho@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>