CBD-6118: Work around a recurring repo problem with --depth=1
There's a very subtle bug when using `repo init --depth=1` in
conjunction with specifying a local directory as the manifests URL.  If
this local manifests directory subsequently gets updated, git sees the
`.repo/manifests` directory as being diverged. This in turn makes `repo`
attempt to rebase `.repo/manifests` onto the new commit, but it does so
in part by referencing `<commit>^1`. This expression can't be resolved
due to the use of `--depth=1`, and so the rebase fails, leading to
errors like

fatal: invalid upstream '932760ddbb27bc296179022bcb2f93e9071c7919^1'

The workaround is to delete `.repo/manifests` entirely before re-running
`repo init`. This is a bit brute-force, but since it's cloning from a
local directory it doesn't take long.

Change-Id: I0d4c652101a4c8c9e8e1ebc5da77db633fdacd38
Reviewed-on: https://review.couchbase.org/c/build-tools/+/222236
Tested-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Blair Watt <blair.watt@couchbase.com>